cloud_function/gcf_with_app_engine/requirements.txt
# Function dependencies, for example:
# package>=version
requests
tensorflow
efficientnet
streamlit
google_api_python_client
protobuf
google-cloud-storage
